requirements.txt
aiogram
sqlalchemy
asyncpg
aiosqlite
marzban
yookassa
pydantic-settings
python-dotenv
faker
fastapi
uvicorn
alembic

